[REFACTOR: test] main의 package 형태와 맞춰줌
