+++ src/test/java/bridge/utils/ValidatorTest.java
@@ -1,6 +1,5 @@
-package bridge;
+package bridge.utils;
 
-import bridge.utils.Validator;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
